requirements.txt
# Core dependencies
streamlit>=1.52.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=6.0.0
pydeck>=0.9.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.2.0

# Hopsworks
hopsworks>=3.0.0
# Note: Install hsfs with: pip install 'hsfs[python]' (quotes needed for zsh)
hsfs[python]>=3.7.0

# Environment
python-dotenv>=1.0.0
